pyproject.toml
[tool.poetry]
name = "hhru-project"
version = "0.1.0"
description = "HH.ru project, search your best work place"
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.6.1"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.9.3"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
